+++ resources/views/categories/index.blade.php
@@ -32,7 +32,7 @@
                             <td>{{ $category->description }}</td>
                             <td>
                                 <a href="/categories/{{ $category->id }}/edit" class="btn btn-warning btn-xs"><i class="fa fa-pencil"></i> Edit</a>
-                                <a href="/categories/delete/{{ $category->id }}" class="btn btn-danger btn-xs"><i class="fa fa-trash"></i> Delete</a>
+                                <a href="/categories/delete/{{ $category->id }}" class="btn btn-danger btn-xs confirm"><i class="fa fa-trash"></i> Delete</a>
                             </td>
                         </tr>
                     @endforeach
