+++ resources/views/permissions/index.blade.php
@@ -32,7 +32,7 @@
                             <td>{{ $permission->label }}</td>
                             <td>
                                 <a href="/permissions/{{ $permission->id }}/edit" class="btn btn-warning btn-xs"><i class="fa fa-pencil"></i> Edit</a>
-                                <a href="/permissions/delete/{{ $permission->id }}" class="btn btn-danger btn-xs"><i class="fa fa-trash"></i> Delete</a>
+                                <a href="/permissions/delete/{{ $permission->id }}" class="btn btn-danger btn-xs confirm"><i class="fa fa-trash"></i> Delete</a>
                             </td>
                         </tr>
                     @endforeach
