+++ resources/views/users/index.blade.php
@@ -42,7 +42,7 @@
                             </td>
                             <td>
                                 <a href="/users/{{ $user->id }}/edit" class="btn btn-warning btn-xs"><i class="fa fa-pencil"></i> Edit</a>
-                                <a href="/users/delete/{{ $user->id }}" class="btn btn-danger btn-xs"><i class="fa fa-trash"></i> Delete</a>
+                                <a href="/users/delete/{{ $user->id }}" class="btn btn-danger btn-xs confirm"><i class="fa fa-trash"></i> Delete</a>
                             </td>
                         </tr>
                     @endforeach
